[Silver III] Title: 주유소, Time: 372 ms, Memory: 38696 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -13,9 +13,9 @@ public class Main {
     }
 
     public static void simulate() {
-        int fee = 0;
+        long fee = 0;
         for (int i = 0; i < n; i++) {
-            int sum = 0;
+            long sum = 0;
             int j = i;
             for (; j < n && station[i] <= station[j]; j++) {
                 sum += road[j];
